changes for inject method
